requirements.txt
Django==5.0.6
dj-database-url==2.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-storages[boto3]==1.14.2
python-dotenv==1.0.1

# Dokumentenverarbeitung für Autofill
docxtpl
pypdf

# KI-Mapping (optional)
openai

# E-Mail Versand (Brevo API)
requests

